refactor(server): use CriteriaBuilder for laboratory listing counts
Replace multi-query approach (findAll + countStationsByLaboratoryIds
joined in service) with single CriteriaBuilder query using correlated
subqueries in LaboratoryRepositoryImpl, matching the institution pattern.

+++ server/src/main/java/com/example/carboncalculator/dto/LaboratoryDTO.java
@@ -8,7 +8,7 @@ public record LaboratoryDTO(
         String name,
         String description,
         boolean active,
-        int configurationCount,
-        int totalStations,
+        long configurationCount,
+        long totalStations,
         OffsetDateTime createdAt) {
 }
